paytm/requirements.txt
# Use Python 3.6
# Main technologies
Django==2.2
djangorestframework==3.9.3
django-rest-knox
coreapi==2.3.3
coreschema==0.0.4
djangorestframework-gis==0.14
django-filter==2.2.0

# Additional support
psycopg2
requests==2.21
Pillow==6.0.0
django-phonenumber-field==2.3.1
phonenumbers==8.10.10
pyyaml==5.1
django-import-export==1.2.0

# AWS
boto3==1.9.138
botocore==1.12.138

#Django-plans
django-plans==0.8.13
django-extensions==2.1.6
mock
tox
suds-jurko
django-rest-multiple-models
django-select-multiple-field

# Documentation
drf-yasg==1.16.0

# Storage
boto==2.49.0
django-storages==1.7.1
sorl-thumbnail==12.5.0
sorl-thumbnail-serializer-field==0.2.1